refactor: centralized erring in response
