--- app/gateway/http/authentication/vo/output/error.go
@@ -1,5 +1,0 @@
-package output
-
-type OutputError struct {
-	Error string `json:"error" example:"error"`
-}
